requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
elevenlabs>=2.0.0
pydantic>=2.5.0
pydantic-core>=2.18.2
pydub>=0.25.0
aiohttp>=3.9.0
